requirement.yxy.txt
python-telegram-bot
requests
pandas
openpyxl